+++ cmd/ani-qw/main.go
@@ -17,7 +17,7 @@ import (
 	"time"
 )
 
-const appVersion = "0.1.14"
+const appVersion = "0.1.15"
 const protocolVersion = 1
 const hostName = "co.aniqw.player"